Extract remaining field types
